Add MapCam, new Render type

+++ Super_Mario_Bros3/Camera.h
@@ -1,12 +1,12 @@
 #pragma once
 class CCamera
 {
-	float cam_x = 0.0f;
-	float cam_y = 0.0f;
+	int cam_x = 0.0f;
+	int cam_y = 0.0f;
 
 public:
 
-	void SetCamPos(float x, float y) { cam_x = x; cam_y = y; }
+	void SetCamPos(int x, int y) { cam_x = x; cam_y = y; }
 
 	int GetCamX()
 	{
